+++ src/abstractsemantics/__init__.py
@@ -1,19 +1,29 @@
 from .registry import (
+    MEMORY_DIGEST_PREDICATE,
+    MemoryRelationDef,
     SemanticsRegistry,
     load_semantics_registry,
     resolve_semantics_registry_path,
 )
 from .schema import (
     KG_ASSERTION_SCHEMA_REF_V0,
+    KG_PREDICATE_ALIAS_MAP_V0,
+    KG_PREDICATE_ALIASES_V0,
     build_kg_assertion_schema_v0,
+    normalize_kg_predicate,
     resolve_schema_ref,
 )
 
 __all__ = [
+    "MEMORY_DIGEST_PREDICATE",
+    "MemoryRelationDef",
     "SemanticsRegistry",
     "load_semantics_registry",
     "resolve_semantics_registry_path",
     "KG_ASSERTION_SCHEMA_REF_V0",
+    "KG_PREDICATE_ALIAS_MAP_V0",
+    "KG_PREDICATE_ALIASES_V0",
     "build_kg_assertion_schema_v0",
+    "normalize_kg_predicate",
     "resolve_schema_ref",
 ]